Stefan

+++ onlineshop-spring-mvc-webapp/src/main/java/com/sci/onlineshop/dao/daoimpl/ProductDaoImpl.java
@@ -38,7 +38,7 @@ public class ProductDaoImpl implements ProductDao {
 	}
 
 	@Override
-	public Product getProductById(int id) {
+	public Product getProductById(long id) {
 		 Session session = sessionFactory.getCurrentSession();
 		 Product product = session.get(Product.class, id);
 		 session.flush();
@@ -57,7 +57,7 @@ public class ProductDaoImpl implements ProductDao {
 	}
 
 	@Override
-	public void deleteProduct(int id) {
+	public void deleteProduct(long id) {
 		Session session = sessionFactory.getCurrentSession();
 		session.delete(getProductById(id));
 		session.flush();
